Remove support for manifest configuration. Unify API for FCM and GCM recievers
